+++ source/FileHandler.cpp
@@ -1,10 +1,10 @@
 #include "FileHandler.h"
 
-FileHandler::FileHandler(std::queue<uint8_t>* buff, int bufferSize) 
+FileHandler::FileHandler(std::queue<uint8_t>* buff, size_t bufferSize) 
     : buff(buff)
     , bufferSize(bufferSize)
      {
-        charBuff = new char [bufferSize];
+        charBuff = new char[bufferSize];
 }
 
 FileHandler::~FileHandler() {
