+++ source/ProcessData.cpp
@@ -3,7 +3,7 @@
 
 ProcessData::ProcessData(std::queue<uint8_t>* buffInput
     , std::queue<uint8_t>* buffOutput
-    , int bufferSize)
+    , size_t bufferSize)
         : buffInput(buffInput)
         , buffOutput(buffOutput)
         , bufferSize(bufferSize) {}
